% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_helpers.R
\name{addBarChartLegendTheme}
\alias{addBarChartLegendTheme}
\title{Helper function to Add Legend and Theme Options to a Bar Chart}
\usage{
addBarChartLegendTheme(font_size, font_family)
}
\arguments{
\item{font_size}{Required, numeric, the font size for the chart.}

\item{font_family}{Required, character, the name of the font family to use for chart.}
}
\value{
A list of ggplot2 guides and theme objects to add to a ggplot2 object.
}
\description{
A function for add all legend and plot theme options to a bar chart, pass as args the
font size and family.
}
