% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggHelpers.R
\name{addBarChartPrePostTheme}
\alias{addBarChartPrePostTheme}
\title{Helper function to Add Theme Options to a Pre-Post Bar Chart}
\usage{
addBarChartPrePostTheme(font_size, font_family)
}
\arguments{
\item{font_size}{Required, numeric, the font size for the chart.}

\item{font_family}{Required, character, the name of the font family to use for chart.}
}
\value{
A list of ggplot2 guides and theme objects to add to a ggplot2 object.
}
\description{
A function for add all plot theme options to a Pre-Post bar chart, pass as args the
font size and family.
}
