% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggHelpers.R
\name{addPlotTag}
\alias{addPlotTag}
\title{Helper function to Add a Plot Tag}
\usage{
addPlotTag(n, font_size, font_family, plot_tag_position = c(-0.01, 1.05))
}
\arguments{
\item{n}{Required, numeric, total n for the sample in the chart}

\item{font_size}{Required, numeric, the font size for the chart.}

\item{font_family}{Required, character, the name of the font family to use for chart.}

\item{plot_tag_position}{Required, character vector of length two that specifies the x and y coordinates that
places the plot tag, defaults to c(-0.01, 1.05).}
}
\value{
A list of ggplot2 labs and theme objects to add to a ggplot2 object.
}
\description{
A function for add a plot tag to the upper left corner of a ggplot with the total n of the sample.
Pass as args the total n, font size and family.
}
