% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowChart.R
\name{arrowChart}
\alias{arrowChart}
\title{Title}
\usage{
arrowChart(df, set_levels, colors_group)
}
\arguments{
\item{df}{A tibble or data frame of categorical data that has a group var to split up the data, e.g. role, gender, education level, etc.}

\item{set_levels}{character vector of labels for the x-axis, usually a response to a set of likert items, needs to match the number of response items in the data.}

\item{colors_group}{character vector of hex codes for colors to associate each group to, e.g. this data has two groups and this function creates an overall group so this function will need a colors_group char vector of three colors- colors need to be in the order you want them associated to the group based on the factor levels for the group variable, last color will be the overall group of "all".}
}
\value{
A ggplot object that plots the items into a arrow bar chart and can be exported.
}
\description{
Title
}
\examples{
items <- dplyr::tibble(
  Pre_Orgs = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Post_Orgs = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  group = c(
    "grad", "undergrad", "grad", "undergrad",
    "grad", "undergrad", "undergrad", "grad", "undergrad"
  )
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")
items <- items \%>\% dplyr::mutate(
  group = factor(group, levels = c("grad", "undergrad"))
)
threeScale_theMark_colors <- c("#79AB53", "#4B9FA6", "#2C2C4F")
arrowChart(items, levels_min_ext, threeScale_theMark_colors)
}
