% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowChartGroup.R
\name{arrowChartGroup}
\alias{arrowChartGroup}
\title{Arrow Chart by Group for Blackstone Research and Evaluation}
\usage{
arrowChartGroup(
  df,
  group,
  scale_labels,
  group_colors = NULL,
  group_levels,
  overall_n = FALSE,
  question_labels = NULL,
  question_order = FALSE,
  font_family = "Arial",
  font_size = 10
)
}
\arguments{
\item{df}{Required, a \link[tibble:tibble-package]{tibble} or data frame of \strong{numeric} data that has items with the prefix of \code{pre_} and \code{post_}; and has a categorical group
variable to split up the data (e.g. role, gender, education level, etc.).}

\item{group}{Required, the name of the grouping variable in a quoted character string, e.g. "role", "gender", "edu_level", etc..}

\item{scale_labels}{Required, a character vector of labels for the response scale, must be in the desired order,
e.g. if you have a 5 item scale of minimal to extensive it should look like this: \code{levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")}.}

\item{group_colors}{Required, A character vector of hex codes for colors to associate
each group supplied in the \code{group_levels} argument. to, e.g. if the data has two groups and this function creates an
overall group so this function will need a 'group_colors' character vector of
three colors. \code{group_colors} need to be in the order you want them associated to
the group based on the factor levels for the group variable. Defaults to BRE custom qualitative palette and
black is always the color for the "Overall" group.}

\item{group_levels}{Required, a character vector of factor levels for the grouping variable, e.g. if the
grouping variable is gender this could be: "male", "female", "non-binary".}

\item{overall_n}{Logical, default is FALSE. If TRUE, returns an overall \emph{n} for all questions that is in the upper left tag of the plot.
If False, adds \emph{n} to each question/item after the respective labels.}

\item{question_labels}{Default is NULL. Takes in a named character vector to both supply labels the questions and sort the order of the questions.
The named character vector should have the new labels as the "name" and the old labels as the "variable" sorted in the
desired order of appearing in the plot, first item will appear at the top of the plot. See examples.}

\item{question_order}{Logical, default is FALSE. If TRUE, the question order will be taken from the user supplied named character vector passed to
question_labels, where the first item will be at the top of the plot and so on. If FALSE, the question order will be the questions with highest
post score average on the top of the plot descending.}

\item{font_family}{Character value to set the font family for all text in the chart, defaults to "Arial".}

\item{font_size}{Numeric value to set the font size in points for all text in the chart, defaults to size 10.}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object that plots the items into a arrow bar chart.
}
\description{
\code{\link[=arrowChartGroup]{arrowChartGroup()}} creates an pre-post arrow chart of group averages and returns
a ggplot object.
}
