% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorHelpers.R
\docType{data}
\name{blackstoneColors}
\alias{blackstoneColors}
\title{blackstone Colors as a Named Vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
blackstoneColors
}
\value{
A named vector of hex colors for Blackstone Research and Evaluation.
}
\description{
A utils function for loading Blackstone Research and Evaluation colors for charts.
}
\keyword{datasets}
