% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{blackstoneExample}
\alias{blackstoneExample}
\title{Get path to \code{blackstone} example data}
\usage{
blackstoneExample(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, all of the example files will be listed.}
}
\description{
\code{blackstone} comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
blackstoneExample()
blackstoneExample("fake_data.csv")
}
