% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bre-package.R
\docType{package}
\name{bre-package}
\alias{bre}
\alias{bre-package}
\title{bre: Blackstone Research and Evaluation Data Analysis and Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of `bre` is to make data manipulation, analysis, and visualization easier and faster for Blackstone Research and Evaluation. `bre` contains functions to create visuals with Blackstone Research and Evaluation branding and as well as common data cleaning, manipulation and analysis tasks for everyone at Blackstone Research and Evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zwcrowley/bre}
  \item \url{https://zwcrowley.github.io/bre/}
  \item Report bugs at \url{https://github.com/zwcrowley/bre/issues}
}

}
\author{
\strong{Maintainer}: Zack Crowley \email{zcrowley@blackstoneevaluation.com}

Other contributors:
\itemize{
  \item Blackstone Research and Evaluation [copyright holder, funder]
}

}
\keyword{internal}
