% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bre_example}
\alias{bre_example}
\title{Get path to \code{bre} example data}
\usage{
bre_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, all of the example files will be listed.}
}
\description{
\code{bre} comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
bre_example()
bre_example("fake_data.csv")
}
