% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorHelpers.R
\docType{data}
\name{breColors}
\alias{breColors}
\title{Helper Functions for colors of visuals for the \code{bre} package}
\format{
An object of class \code{character} of length 4.
}
\usage{
breColors
}
\value{
A named vector of hex colors for Blackstone Research and Evaluation.
}
\description{
A utils function for loading Blackstone Research and Evaluation colors for charts.
}
\details{
BRE Colors as a Named Vector
}
\keyword{datasets}
