% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorHelpers.R
\name{customCols}
\alias{customCols}
\title{Helper function that makes selections from a color palette.}
\usage{
customCols(pal = qualColors(), cols = NULL)
}
\arguments{
\item{pal}{Required, the named color palette as a named vector of color hex codes, defaults to \code{qualColors}.}

\item{cols}{Required, a character vector of names or numbered position to use to select
from the named vector of color hex codes , defaults to \code{NULL} and returns the full palette.}
}
\value{
a named character vector of color hex codes.
}
\description{
A function return hex color codes from a color palette by name or numbered position.
}
\keyword{internal}
