% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divBarChart.R
\name{divBarChart}
\alias{divBarChart}
\title{Diverging and Stacked Bar Chart for The Mark USA, Inc.}
\usage{
divBarChart(
  df,
  scale_labels,
  percent_label = TRUE,
  question_order = NULL,
  question_labels = NULL,
  width = NULL
)
}
\arguments{
\item{df}{A tibble/data frame of survey items that are categorical/character
variables, in 5 point scales and pre-post, that will be inserted into a
diverging bar chart with The Mark USA branding.}

\item{scale_labels}{Required, a character vector of 5 levels to set the scale for the plot.}

\item{percent_label}{Default is TRUE. If FALSE, labels the bars with the number of answers per response.}

\item{question_order}{Takes in a character vector to sort the order of the questions. Default is NULL.}

\item{question_labels}{Takes in a character vector to label the questions. Needs to be in same order as question_order. Default is NULL.}

\item{width}{Input a value between 0.3 and 0.8 to set the thickness of the bars. Default is NULL.}
}
\value{
A ggplot object that plots the items into a diverging and stacked bar
chart as a ggplot object. The chart is sorted by the highest positive
valence post items at the top, or by user supplied question order.
}
\description{
Diverging and Stacked Bar Chart for The Mark USA, Inc.
}
\examples{
items <- dplyr::tibble(
  Pre_Organization = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Post_Organization = dplyr::if_else(Pre_Organization < 5, Pre_Organization + 1, Pre_Organization),
  Pre_Source = c(2, 2, 3, 5, 4, 3, 2, 1, 2),
  Post_Source = dplyr::if_else(Pre_Source < 4, Pre_Source + 2, Pre_Source),
  Pre_Publish = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  Post_Publish = Pre_Publish + 2,
  Pre_Write = c(2, 2, 2, 3, 3, 3, 4, 4, 4),
  Post_Write = Pre_Write + 1,
  Pre_Research = c(1, 1, 2, 2, 3, 3, 4, 4, 4),
  Post_Research = Pre_Research + 1
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")

# Recode the numeric to factor variables using the levels from levels_min_ext:
cat_items <- TheMarkUSA::recodeCat(items, levels_min_ext)

# Select the factor variables:
cat_items <- cat_items \%>\% dplyr::select(dplyr::where(is.factor))

divBarChart(cat_items, levels_min_ext)
}
