% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{getHeaders}
\alias{getHeaders}
\title{Create Codebooks and Read in Data with New Names from SurveyMonkey Data}
\usage{
getHeaders(file_path)
}
\arguments{
\item{file_path}{Required, A file path with extension of .csv or .xlsx.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the data in 3 columns: \code{question_name}, \code{full_text}, and \code{full_question_text}.
\code{question_name} is the first header row, \code{full_text} is the second header row, and \code{full_question_text} is the combination of the two headers.
}
\description{
\code{\link[=getHeaders]{getHeaders()}}, \code{\link[=createCodebook]{createCodebook()}}, and \code{\link[=readRenameData]{readRenameData()}} provide tools for
dealing with survey data from SurveyMonkey.
\code{\link[=getHeaders]{getHeaders()}} is a helper function that turns the two-headers from SurveyMonkey
data into a three column tibble where the third column \code{full_question_text}
is the combination of the two headers.
\code{\link[=createCodebook]{createCodebook()}} returns the three columns from \code{getHeaders()} and adds two more
columns, \code{variable_name} which is a cleaned up version of \code{full_question_text} and
\code{position} which is the column position number for each \code{variable_name}.
\code{\link[=readRenameData]{readRenameData()}} takes in a file path and the codebook created from  \code{createCodebook()}
and reads in the data with the updated variable names.

Convert two row headers from SurveyMonkey data into long from data

\code{getHeaders()} is a helper function to get headers into a long form tibble.
}
