% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{qualColors}
\alias{qualColors}
\title{BRE Qualitative Colors as a Vector}
\usage{
qualColors(add_names = FALSE)
}
\arguments{
\item{add_names}{Required, logical, if FALSE returns a vector of hex color codes, if TRUE returns a named vector with color names
that can be used to select from.}
}
\value{
A vector or named vector of hex colors for Qualitative color scale for Blackstone Research and Evaluation.
}
\description{
A utils function for loading Qualitative color scale Blackstone Research and Evaluation colors for charts.
}
