% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorHelpers.R
\name{qualFillColors}
\alias{qualFillColors}
\title{Helper function to create qualitative colors from the \code{Okabe-Ito} palette.}
\usage{
qualFillColors(n_colors = 11, rev_colors = FALSE)
}
\arguments{
\item{n_colors}{Required, supply a non-negative integer that is the desired the number of color hex codes to return.}

\item{rev_colors}{Logical, defaults to FALSE, if true returns a reverse color codes where the darkest color comes first.}
}
\value{
a character vector of hex color codes the length of \code{n_colors}.
}
\description{
A function to create a custom qualitative colors palette from the \code{Okabe-Ito} palette, reversed if rev_colors is set to TRUE.
Drops black and yellow for use with all Blackstone charts and adds three other colors: "#440154FF", "#283251", and "#999933".
}
\examples{
# Returns the full color palette:
qualFillColors()

# Returns the first 5 colors in the palette:
qualFillColors(n_colors = 5)

# Returns the first 5 colors in the palette reversed:
qualFillColors(n_colors = 5, rev_colors = TRUE)
}
