% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readRenameData}
\alias{readRenameData}
\title{Import SurveyMonkey data and create new variable names}
\usage{
readRenameData(file_path, codebook)
}
\arguments{
\item{file_path}{Required, A file path with extension of .csv or .xlsx.}

\item{codebook}{Required, A \link[tibble:tibble-package]{tibble} created by using the function \code{\link[=createCodebook]{createCodebook()}} that has the column named \code{variable_name}
that will be the new names of the data imported into R.}
}
\value{
A \link[tibble:tibble-package]{tibble} of SurveyMonkey data with new variable names.
}
\description{
\code{readRenameData()} Reads in SurveyMonkey data with new variable names taken from the codebook \link[tibble:tibble-package]{tibble} column named \code{variable_name}
created by using the function \code{\link[=createCodebook]{createCodebook()}}.
}
