% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeFiveCat.R
\name{recodeFiveCat}
\alias{recodeFiveCat}
\title{Title}
\usage{
recodeFiveCat(df, set_5_levels)
}
\arguments{
\item{df}{A tibble/dataframe of survey items that are categorical/character variables, in 5 point scales and pre-post, that will be inserted into a stacked bar chart with The Mark USA branding.}

\item{set_5_levels}{character vector of 5 levels to set the numeric variables to factor/categorical valuse}
}
\value{
new tibble/dataframe with the original numeric variables along with new variables mapped to categorical values from set_5_levels character vector.
}
\description{
Title
}
\examples{
items <- dplyr::tibble(Pre_Orgs = c(1,2,3,4,5,4,3,2,1), Post_Orgs = c(1,2,3,4,5,4,3,2,1))
levels_min_ext <- c('Minimal', 'Slight', 'Moderate', 'Good', 'Extensive')
recodeFiveCat(items ,levels_min_ext)
}
