% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarChart.R
\name{stackedBarChart}
\alias{stackedBarChart}
\title{Stacked Bar Chart for The Mark Reporting}
\usage{
stackedBarChart(df)
}
\arguments{
\item{df}{A tibble/dataframe of survey items that are categorical/character variables, in 5 point scales and pre-post, that will be inserted into a stacked bar chart with The Mark USA branding.}
}
\value{
A ggplot object that plots the items into a stacked bar chart and can be exported.
}
\description{
Stacked Bar Chart for The Mark Reporting
}
\examples{
items <- tibble(cat_Pre_Sources = c("Good","Moderate","Minimal","Slight","Slight","Moderate","Good"),
              cat_Post_Sources = c("Good","Good","Minimal","Moderate","Moderate","Good","Extensive"),
              cat_Pre_Orgs = c("Good","Moderate","Minimal","Slight","Slight","Moderate","Moderate"),
              cat_Post_Orgs = c("Good","Moderate","Minimal","Moderate","Moderate","Good","Extensive"))
stackedBarChart(items)
}
