% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFlextable.R
\name{testFlextable}
\alias{testFlextable}
\title{Create Flextable in Blackstone Research and Evaluation Branding}
\usage{
testFlextable(data)
}
\arguments{
\item{data}{Required, A \link[tibble:tibble-package]{tibble}/data frame of summary/descriptice or
test statistics.}
}
\value{
A nicely formatted table that is a \code{\link[=flextable]{flextable()}} object.
}
\description{
Create Flextable in Blackstone Research and Evaluation Branding
}
\examples{
# Example data:
data <- dplyr::tibble(
  Organization = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Source = c(2, 2, 3, 5, 4, 3, 2, 1, 2),
  Publish = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  Write = c(2, 2, 2, 3, 3, 3, 4, 4, 4),
  Research = c(1, 1, 2, 2, 3, 3, 4, 4, 4),
  group = factor(c(
      "grad", "undergrad", "grad", "undergrad", "grad",
      "undergrad", "undergrad", "grad", "undergrad"
                    ), levels = c("grad", "undergrad")
    )
)
# Summarise the data into mean and sd for each numeric column, grouped by "group":
data <- data \%>\% dplyr::group_by(group) \%>\%   dplyr::summarise(dplyr::across(
                                   .cols = dplyr::where(is.numeric),
                                   .fns =  list(Mean = mean, SD = sd),
                                   .names = "{col}_{fn}"
                                                  )) \%>\%
             dplyr::ungroup() \%>\%
             tidyr::pivot_longer(cols = -group,
                                 names_pattern = "([^_]+)_(.*)",
                                 names_to = c("variable", "stat"),
                                 values_to = "value",
                                 values_drop_na = TRUE) \%>\%
             tidyr::pivot_wider(names_from = stat, values_from = value)

# Make a nice table with the function:
data \%>\% testFlextable()
}
