% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheMarkUSA-package.R
\docType{package}
\name{TheMarkUSA-package}
\alias{TheMarkUSA}
\alias{TheMarkUSA-package}
\title{TheMarkUSA: Helper Functions for The Mark USA, INC}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of TheMarkUSA is to make data cleaning and the creation of visualizations easier and faster for The Mark USA, INC. TheMarkUSA contains functions to create visuals with The Mark USA branding and helper functions for common data cleaning and manipulation tasks for use for everyone at The Mark USA, INC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zcrowleythemark.github.io/TheMarkUSA/}
  \item \url{https://github.com/zcrowleyTheMark/themark}
  \item Report bugs at \url{https://github.com/zcrowleyTheMark/themark/issues}
}

}
\author{
\strong{Maintainer}: Zack Crowley \email{zcrowley@themarkusa.com}

Other contributors:
\itemize{
  \item The Mark USA, INC. [copyright holder, funder]
}

}
\keyword{internal}
