% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowChart.R
\name{arrowChart}
\alias{arrowChart}
\title{Arrow Chart for The Mark USA, Inc.}
\usage{
arrowChart(df, scale_labels, group_colors)
}
\arguments{
\item{df}{A tibble or data frame of numeric data that has a categorical group
variable to split up the data, e.g. role, gender, education level, etc.}

\item{scale_labels}{character vector of labels for the x-axis, usually a
response to a set of likert items, needs to match the number of response
items in the data.}

\item{group_colors}{character vector of hex codes for colors to associate
each group to, e.g. this data has two groups and this function creates an
overall group so this function will need a 'group_colors' character vector of
three colors. 'group_colors' need to be in the order you want them associated to
the group based on the factor levels for the group variable, last color
will be the overall group of "all".}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object that plots the items into a arrow bar chart.
}
\description{
\code{arrowChart()} creates an pre-post arrow chart of group averages and returns
a ggplot object.
}
\examples{
items <- dplyr::tibble(
  Pre_Organization = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Post_Organization = dplyr::if_else(Pre_Organization < 5, Pre_Organization + 1, Pre_Organization),
  Pre_Source = c(2, 2, 3, 5, 4, 3, 2, 1, 2),
  Post_Source = dplyr::if_else(Pre_Source < 4, Pre_Source + 2, Pre_Source),
  Pre_Publish = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  Post_Publish = Pre_Publish + 2,
  Pre_Write = c(2, 2, 2, 3, 3, 3, 4, 4, 4),
  Post_Write = Pre_Write + 1,
  Pre_Research = c(1, 1, 2, 2, 3, 3, 4, 4, 4),
  Post_Research = Pre_Research + 1,
  group = c(
    "grad", "grad", "grad", "grad",
    "undergrad", "undergrad", "undergrad", "undergrad", "undergrad"
  )
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")
items <- items \%>\% dplyr::mutate(
  group = factor(group, levels = c("grad", "undergrad"))
)
threeScale_theMark_colors <- c("#79AB53", "#4B9FA6", "#2C2C4F")
arrowChart(df = items, scale_labels = levels_min_ext, group_colors = threeScale_theMark_colors)
}
