% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bre-package.R
\docType{package}
\name{bre-package}
\alias{bre}
\alias{bre-package}
\title{bre: Helper Functions for Blackstone Research and Evaluation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of `bre` is to make data cleaning and the creation of visualizations easier and faster for Blackstone Research and Evaluation. `bre` contains functions to create visuals with Blackstone Research and Evaluation branding and helper functions for common data cleaning and manipulation tasks for everyone at Blackstone Research and Evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zcrowleythemark.github.io/bre/}
  \item \url{https://github.com/zcrowleyTheMark/themark}
  \item Report bugs at \url{https://github.com/zcrowleyTheMark/themark/issues}
}

}
\author{
\strong{Maintainer}: Zack Crowley \email{zcrowley@blackstoneevaluation.com}

Other contributors:
\itemize{
  \item Blackstone Research and Evaluation [copyright holder, funder]
}

}
\keyword{internal}
