% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{createCodebook}
\alias{createCodebook}
\title{Create a codebook for SurveyMonkey data}
\usage{
createCodebook(file_path)
}
\arguments{
\item{file_path}{Required, A file path with extension of .csv or .xlsx.}
}
\value{
A \link[tibble:tibble-package]{tibble} with 5 columns: \code{question_name}, \code{full_text}, \code{full_question_text}, \code{variable_name} and \code{position}.
\code{question_name} is the first header row, \code{full_text} is the second header row, \code{full_question_text} is the combination of the two headers,
\code{variable_name} is a cleaned up version for \code{full_question_text} and will be the column to edit to change the column names later on to more
meaningful and shorter names, and \code{position} is the column number for each \code{variable_name}.
}
\description{
\code{createCodebook()} creates a partial codebook that can be edited to create useful variable names from SurveyMonkey data.
It returns a long form tibble to use as a code book for the SurveyMonkey data, it returns 5 columns described below and \code{variable_name}
is the column that will be used in the function \code{\link[=readRenameData]{readRenameData()}} to rename the variable names in the data to be imported into R.
}
