% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSumm.R
\name{dataSumm}
\alias{dataSumm}
\title{Creates a data frame that is a summary table of counts and percentages}
\usage{
dataSumm(var, na.rm = TRUE)
}
\arguments{
\item{var}{A column selected from a \link[tibble:tibble-package]{tibble}/data frame that is a categorical/factor variable to that to be summarized into a table.}

\item{na.rm}{Drops NA values. Defaults to TRUE.}
}
\value{
a \link[tibble:tibble-package]{tibble} with the data in 5 columns: item, response, n_answers, percent_answers and percent_answers_label.
Item is the name of the original item, Response is all of the categorical responses possible for the item. n_answers is the count of each response,
percent_answers is the percentage of each response and percent_answers_label is a character variable of percentage labelled with percent sign for use as a label.
}
\description{
Creates a data frame that is a summary table of counts and percentages
}
\examples{
data <- dplyr::tibble(
  role = c(
    "Faculty", "Postdoc", "Undergraduate student", "Graduate student",
    "Graduate student", "Postdoc", "Postdoc", "Faculty",
    "Faculty", "Graduate student", "Graduate student", "Postdoc",
    "Faculty", "Faculty", "Faculty", "Faculty", "Faculty", "Graduate student",
    "Undergraduate student", "Undergraduate student", "NA", "NA"
  )
)

data \%>\%
  dplyr::select(role) \%>\%
  dataSumm()

# Includes NA values:
data \%>\%
  dplyr::select(role) \%>\%
  dataSumm(na.rm = TRUE)
}
