% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divBarChart.R
\name{divBarChart}
\alias{divBarChart}
\title{Diverging and Stacked Bar Chart for The Mark USA, Inc.}
\usage{
divBarChart(
  df,
  scale_labels,
  overall_n = FALSE,
  percent_label = TRUE,
  question_labels = NULL,
  question_order = FALSE,
  width = NULL
)
}
\arguments{
\item{df}{Required, A \link[tibble:tibble-package]{tibble}/data frame of survey items that are categorical/character
variables, in 3 to 7 point scales, that will be inserted into a diverging bar chart with The Mark USA branding.}

\item{scale_labels}{Required, a character vector of levels to set the scale for the plot,
accepts a character vector of 3 to 7 items.}

\item{overall_n}{Logical, default is FALSE. If TRUE, returns an overall n for all questions that is in the upper left tag of the plot.}

\item{percent_label}{Logical, default is TRUE. If FALSE, labels the bars with the number of answers per response.}

\item{question_labels}{Default is NULL. Takes in a named character vector to both supply labels the questions and sort the order of the questions.
The named character vector should have the new labels as the "name" and the old labels as the "variable" sorted in the
desired order of appearing in the plot, first item will appear at the top of the plot. See examples.}

\item{question_order}{Logical, default is FALSE. If TRUE, the question order will be taken from the user supplied named character vector passed to
question_labels, where the first item will be at the top of the plot and so on. If FALSE, the question order will be the questions with highest
positive valenced response options on the top of the plot descending.}

\item{width}{Input a value between 0.3 and 0.8 to set the thickness of the bars. Default is NULL.}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object that plots the items into a diverging bar chart and can be exported.
}
\description{
Diverging and Stacked Bar Chart for The Mark USA, Inc.
}
