% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divBarChart.R
\name{divBarChart}
\alias{divBarChart}
\title{Diverging and Stacked Bar Chart for The Mark}
\usage{
divBarChart(df, set_5_levels)
}
\arguments{
\item{df}{A tibble/data frame of survey items that are categorical/character variables, in 5 point scales and pre-post, that will be inserted into a stacked bar chart with The Mark USA branding.}

\item{set_5_levels}{character vector of 5 levels to set the scale for the plot}
}
\value{
A ggplot object that plots the items into a diverging and stacked bar chart and can be exported.
}
\description{
Diverging and Stacked Bar Chart for The Mark
}
\examples{
items <- dplyr::tibble(
  cat_Pre_Sources =
    c("Good", "Moderate", "Minimal", "Slight", "Slight", "Moderate", "Good"),
  cat_Post_Sources =
    c("Good", "Good", "Minimal", "Moderate", "Moderate", "Good", "Extensive"),
  cat_Pre_Orgs =
    c("Good", "Moderate", "Minimal", "Slight", "Slight", "Moderate", "Moderate"),
  cat_Post_Orgs =
    c("Good", "Moderate", "Minimal", "Moderate", "Moderate", "Good", "Extensive")
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")
divBarChart(items, levels_min_ext)
}
