% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{divFillColors}
\alias{divFillColors}
\title{Helper to create a sequential color scale using \verb{Blue-Red 3} that is reversed.}
\usage{
divFillColors(n_colors)
}
\arguments{
\item{n_colors}{Required, the number of color hex codes to return.}
}
\value{
a character vector hex color codes the length of \code{n_colors}from the \verb{Blue-Red 3} palette from the package \code{colorspace}.
}
\description{
A function to create a sequential color scale using \verb{Blue-Red 3} that is reversed and slightly darkened.
}
