% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{getHeaders}
\alias{getHeaders}
\title{Convert two row headers from SurveyMonkey data into long from data}
\usage{
getHeaders(file_path)
}
\arguments{
\item{file_path}{Required, A file path with extension of .csv or .xlsx.}
}
\value{
A \link[tibble:tibble-package]{tibble} with 3 columns: \code{question_name}, \code{full_text}, and \code{full_question_text}.
\code{question_name} is the first header row, \code{full_text} is the second header row, and \code{full_question_text} is the combination of the two headers.
}
\description{
\code{getHeaders()} is a helper function to get headers into a long form tibble.
}
