% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{getHeaders}
\alias{getHeaders}
\title{Convert two row headers from SurveyMonkey data into long from data}
\usage{
getHeaders(file_path)
}
\arguments{
\item{file_path}{Required, A file path with extension of .csv or .xlsx.}
}
\value{
A \link[tibble:tibble-package]{tibble} with 3 columns: \code{header_1}, \code{header_2}, and \code{combined_header}.
\code{header_1} is the first header row, \code{header_2} is the second header row, and \code{combined_header} is the combination of the two headers.
}
\description{
\code{getHeaders()} is a helper function to get headers into a long form tibble.
}
