% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horzBarChart.R
\name{horzBarChart}
\alias{horzBarChart}
\title{Horizontal Bar Chart for The Mark USA, Inc.}
\usage{
horzBarChart(df, scale_colors, width = NULL)
}
\arguments{
\item{df}{Required, a tibble or data frame that has been pre-processed with dataSumm().}

\item{scale_colors}{Required, a character vector of the colors for the scale items.}

\item{width}{Input a value between 0.3 and 0.8 to set the thickness of the bars. Default is NULL.}
}
\value{
A ggplot object that plots the items into a stacked bar chart and can be exported.
}
\description{
Horizontal Bar Chart for The Mark USA, Inc.
}
\examples{
data <- dplyr::tibble(
  role = c(
    "Faculty", "Postdoc", "Undergraduate student", "Graduate student",
    "Graduate student", "Postdoc", "Postdoc", "Faculty",
    "Faculty", "Graduate student", "Graduate student", "Postdoc",
    "Faculty", "Faculty", "Faculty", "Faculty", "Faculty", "Graduate student",
    "Undergraduate student", "Undergraduate student"
  )
)

role_summ <- data \%>\%
  dplyr::select(role) \%>\%
  TheMarkUSA::dataSumm()

role_color <- c("#2C2C4F", "#4B9FA6", "#79AB53", "#767171")

horzBarChart(df = role_summ, scale_colors = role_color, width = 0.6)
}
