% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_levels.R
\docType{data}
\name{levelsEnv}
\alias{levelsEnv}
\title{Levels Environment and home to all scale_labels vectors:}
\format{
An object of class \code{environment} of length 3.
}
\usage{
levelsEnv
}
\description{
Levels Environment and home to all scale_labels vectors:
}
\keyword{datasets}
