% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\docType{data}
\name{qualColors}
\alias{qualColors}
\title{BRE Qualitative Colors as a Named Vector}
\format{
An object of class \code{character} of length 11.
}
\usage{
qualColors
}
\value{
A named vector of hex colors for Qualitative color scale for Blackstone Research and Evaluation.
}
\description{
A utils function for loading Qualitative color scale Blackstone Research and Evaluation colors for charts.
}
\keyword{datasets}
