% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{qualColors}
\alias{qualColors}
\title{BRE Qualitative Colors as a Vector}
\usage{
qualColors(add_names = FALSE)
}
\arguments{
\item{add_names}{Required, logical, if FALSE returns a vector of hex color codes, if TRUE returns a named vector with color names
that can be used to select from.}
}
\value{
A vector or named vector of hex colors for Qualitative color scale for Blackstone Research and Evaluation.
}
\description{
A utils function for loading Qualitative color scale Blackstone Research and Evaluation colors for charts.
}
\examples{
# Full color palette with names:
qualColors(add_names = TRUE)

# function to show color, names and hex codes as visual:
show_colors2 <- function(colors) {
    labels_color <- purrr::map_chr(seq_along(colors),
                                   \(x) paste0("'",names(colors)[x], "': ", colors[x]))
    labels_text_color <- labelColorMaker(colors = colors)
    ggplot2::ggplot(data.frame(id = rev(seq_along(colors)), color = rev(colors))) +
        ggplot2::geom_tile(ggplot2::aes(1, id, fill = rev(color))) +
        ggplot2::geom_text(ggplot2::aes(1, id, label = labels_color), color = labels_text_color) +
        ggplot2::scale_fill_identity() +
        ggplot2::scale_color_identity() +
        ggplot2::theme_void()
}
show_colors2(colors = qualColors(add_names = TRUE))
}
