% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{qualFillColors}
\alias{qualFillColors}
\title{Helper function to create qualitative colors: either default to \code{viridis} or \code{Okabe-Ito}.}
\usage{
qualFillColors(n_colors, pal = "viridis", rev_colors = FALSE)
}
\arguments{
\item{n_colors}{Required, the number of color hex codes to return.}

\item{pal}{Required, the name of the palette to use, defaults to \code{viridis} or can use \code{Okabe-Ito}.}

\item{rev_colors}{Logical, defaults to FALSE, if true returns a reverse color codes where the darkest color comes first.}
}
\value{
a character vector of hex color codes the length of \code{n_colors}.
}
\description{
A function to create qualitative colors: pal argument either default to \code{viridis} or \code{Okabe-Ito}, reversed if rev_colors is set to TRUE.
}
