% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeCat.R
\name{recodeCat}
\alias{recodeCat}
\title{Recode Numeric Variables to Factors Variables}
\usage{
recodeCat(df, scale_labels)
}
\arguments{
\item{df}{A tibble/data frame of survey items that are numeric
variables that need to be converted into factor variables.}

\item{scale_labels}{character vector of levels to set the numeric variables
to factor/categorical values}
}
\value{
a \link[tibble:tibble-package]{tibble} with the original numeric variables along with
new variables mapped to categorical values from scale_labels character
vector.
}
\description{
Recode Numeric Variables to Factors Variables
}
\examples{
items <- dplyr::tibble(
  Pre_Organization = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Post_Organization = dplyr::if_else(Pre_Organization < 5, Pre_Organization + 1, Pre_Organization),
  Pre_Source = c(2, 2, 3, 5, 4, 3, 2, 1, 2),
  Post_Source = dplyr::if_else(Pre_Source < 4, Pre_Source + 2, Pre_Source),
  Pre_Publish = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  Post_Publish = Pre_Publish + 2,
  Pre_Write = c(2, 2, 2, 3, 3, 3, 4, 4, 4),
  Post_Write = Pre_Write + 1,
  Pre_Research = c(1, 1, 2, 2, 3, 3, 4, 4, 4),
  Post_Research = Pre_Research + 1
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")
recodeCat(items, levels_min_ext)
}
