% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seqFillColors}
\alias{seqFillColors}
\title{Helper to create a sequential color scale using \verb{Blues 3} that is reversed.}
\usage{
seqFillColors(n_colors)
}
\arguments{
\item{n_colors}{Required, the number of color hex codes to return.}
}
\value{
a character vector hex color codes the length of \code{n_colors} from the \verb{Blues 3} palette.
}
\description{
A function to create a sequential color scale using \verb{Blues 3}, that is reversed.
}
