% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarChart.R
\name{stackedBarChart}
\alias{stackedBarChart}
\title{Stacked Bar Chart for The Mark USA, Inc.}
\usage{
stackedBarChart(df, set_5_levels)
}
\arguments{
\item{df}{A tibble/data frame of survey items that are categorical/character
variables, in 5 point scales and pre-post, that will be inserted into a
stacked bar chart with The Mark USA branding.}

\item{set_5_levels}{character vector of 5 levels to set the scale for the
plot}
}
\value{
A ggplot object that plots the items into a stacked bar chart and can
be exported.
}
\description{
Stacked Bar Chart for The Mark USA, Inc.
}
\examples{
items <- dplyr::tibble(
  Pre_Organization = c(1, 2, 3, 4, 5, 4, 3, 2, 1),
  Post_Organization = dplyr::if_else(Pre_Organization < 5, Pre_Organization + 1, Pre_Organization),
  Pre_Source = c(2, 2, 3, 5, 4, 3, 2, 1, 2),
  Post_Source = dplyr::if_else(Pre_Source < 4, Pre_Source + 2, Pre_Source),
  Pre_Publish = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  Post_Publish = Pre_Publish + 2,
  Pre_Write = c(2, 2, 2, 3, 3, 3, 4, 4, 4),
  Post_Write = Pre_Write + 1,
  Pre_Research = c(1, 1, 2, 2, 3, 3, 4, 4, 4),
  Post_Research = Pre_Research + 1
)
levels_min_ext <- c("Minimal", "Slight", "Moderate", "Good", "Extensive")

# Recode the numeric to factor variables using the levels from levels_min_ext:
cat_items<- TheMarkUSA::recodeFiveCat(items, levels_min_ext)

# Select the factor variables:
cat_items <- cat_items \%>\% dplyr::select(dplyr::where(is.factor))

# Pass the factor variables and the levels to 'stackedBarChart()':
stackedBarChart(cat_items, levels_min_ext)
}
