% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheMarkUSA-package.R
\docType{package}
\name{TheMarkUSA-package}
\alias{TheMarkUSA}
\alias{TheMarkUSA-package}
\title{TheMarkUSA: Helper Functions for The Mark, USA Inc}
\description{
Functions that create visuals with The Mark USA, Inc. branding and helper functions for common data cleaning and manipulation tasks for use for everyone at The Mark, USA Inc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zcrowleyTheMark/themark}
  \item Report bugs at \url{https://github.com/zcrowleyTheMark/themark/issues}
}

}
\author{
\strong{Maintainer}: Zack Crowley \email{zcrowley@themarkusa.com}

Other contributors:
\itemize{
  \item The Mark, USA Inc. [copyright holder, funder]
}

}
\keyword{internal}
